+++ src/main/java/com/almostreliable/kubeio/event/VatReagentModificationEvent.java
@@ -2,39 +2,22 @@ package com.almostreliable.kubeio.event;
 
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.crafting.Ingredient;
 
-import com.mojang.datafixers.util.Either;
-import dev.latvian.mods.kubejs.event.KubeEvent;
 import dev.latvian.mods.kubejs.item.ItemPredicate;
 import dev.latvian.mods.kubejs.script.data.VirtualDataMapFile;
+import dev.latvian.mods.rhino.util.HideFromJS;
 
 import java.util.Map;
-import java.util.stream.Stream;
 
-public record VatReagentModificationEvent(
-    VirtualDataMapFile<Item, Map<TagKey<Item>, Double>> dataMap
-) implements KubeEvent {
+public class VatReagentModificationEvent extends SimpleDataMapEvent<Map<TagKey<Item>, Double>> {
+
+    @HideFromJS
+    public VatReagentModificationEvent(VirtualDataMapFile<Item, Map<TagKey<Item>, Double>> dataMap) {
+        super(dataMap);
+    }
 
     public void add(ItemPredicate item, TagKey<Item> tag, double modifier) {
-        dissolve(item)
-            .ifLeft(t -> dataMap.addTag(t, Map.of(tag, modifier)))
-            .ifRight(items -> items.forEach(i -> dataMap.add(i, Map.of(tag, modifier))));
-    }
-
-    public void remove(ItemPredicate item) {
-        dissolve(item)
-            .ifLeft(dataMap::removeTag)
-            .ifRight(s -> s.forEach(dataMap::remove));
-    }
-
-    public void clear() {
-        dataMap.clear();
-    }
-
-    private static Either<TagKey<Item>, Stream<Item>> dissolve(ItemPredicate filter) {
-        var tag = filter instanceof Ingredient ingredient ? ingredient.kjs$getTagKey() : null;
-        if (tag != null) return Either.left(tag);
-        return Either.right(filter.kjs$getItemStream());
+        var data = Map.of(tag, modifier);
+        add(item, data);
     }
 }