+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 ## Unreleased
-- /
+
+- updated to EnderIO 8.2.0 ([#24](https://github.com/AlmostReliable/kubejs-enderio/issues/24))
+- changed Grinding Ball event from startup to server type
 
 ## [0.12.0] - 2026-02-18